+++ tests/regression/redis/scripting.py
@@ -16,7 +16,7 @@ binary_set = lambda x : set(to_charlist(x))
 class test_script(redis.RedisScript):
     script = (redis.read_lua_file('commands.utils'),
               '''\
-js = cjson.decode(ARGV[1])
+local js = cjson.decode(ARGV[1])
 return cjson.encode(js)''')
     
     def callback(self, request, result, args, **options):