+++ CHANGELOG.rst
@@ -54,7 +54,7 @@ Ver. 0.7c3 - 2012 May 02
 * Added :mod:`stdnet.utils.path`.
 * Added a Lua test suite for testing stand alone scripts. Requires lunatest_.
 * PEP 386-compliant version number.
-* **572 regression tests** with **90%** coverage.
+* **573 regression tests** with **90%** coverage.
 
 .. _vers06: